+++ databento/live/session.py
@@ -328,7 +328,8 @@ class Session:
                 return
             if self._transport.can_write_eof():
                 self._loop.call_soon_threadsafe(self._transport.write_eof)
-            self._loop.call_soon_threadsafe(self._transport.close)
+            else:
+                self._loop.call_soon_threadsafe(self._transport.close)
 
     def subscribe(
         self,