+++ tests/test_live_client.py
@@ -427,27 +427,13 @@ async def test_live_subscribe_large_symbol_list(
         symbols=large_symbol_list,
     )
 
-    first_message = mock_live_server.get_message_of_type(
-        gateway.SubscriptionRequest,
-        timeout=1,
-    ).symbols.split(",")
-
-    second_message = mock_live_server.get_message_of_type(
-        gateway.SubscriptionRequest,
-        timeout=1,
-    ).symbols.split(",")
-
-    third_message = mock_live_server.get_message_of_type(
-        gateway.SubscriptionRequest,
-        timeout=1,
-    ).symbols.split(",")
-
-    fourth_message = mock_live_server.get_message_of_type(
-        gateway.SubscriptionRequest,
-        timeout=1,
-    ).symbols.split(",")
-
-    reconstructed = first_message + second_message + third_message + fourth_message
+    reconstructed: list[str] = []
+    for _ in range(8):
+        message = mock_live_server.get_message_of_type(
+            gateway.SubscriptionRequest,
+            timeout=1,
+        ).symbols.split(",")
+        reconstructed.extend(message)
 
     # Assert
     assert reconstructed == large_symbol_list